closeAlert {shinyalert}	R Documentation

Close a shinyalert popup message

Description

Close a shinyalert popup message

Usage

closeAlert(num = 0, id = NULL)

Arguments

num
Number of popup messages to close. If set to 0 (default) then all messages are closed. This is only useful if you have multiple popups queued up.

id
To close a specific popup, use the ID returned by shinyalert. Note that if id is specified, then num is ignored.

[Package shinyalert version 3.1.0 Index]