runExample {shinyalert}	R Documentation

Run shinyalert example

Description

Launch an example Shiny app that shows how easy it is to create modals with shinyalert.

The demo app is also available online to experiment with.

Usage

runExample()

[Package shinyalert version 3.1.0 Index]